2005-10-11 00:45 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/xharbour.spec   * xharbour/bin/pack_src.sh     * updated for recent directory modifications
  * xharbour/contrib/rdd_ads/ads1.c
    * cleaned two warnings

  * xharbour/include/hbset.h
  * xharbour/source/rtl/set.c
  * xharbour/source/vm/hvm.c
  * xharbour/source/vm/thread.c
    % optimized MT pcode counter
    % moved default value checking from HVM loop to SET init code
    ! fixed BKG counter checking
    * changed HB_SET_BACKGROUNDTASKS from ULONG to int - due to
      mistake in set() it was always in 'int' range so this modification
      does not effectively change any working code behavior

+++ xharbour/include/hbset.h
@@ -1,5 +1,5 @@
 /*
- * $Id: hbset.h,v 1.32 2005/03/06 23:43:20 andijahja Exp $
+ * $Id: hbset.h,v 1.33 2005/09/11 19:39:32 druzus Exp $
  */
 
 /*
@@ -220,7 +220,7 @@ typedef struct
    FHANDLE hb_set_winhan;
    char *  hb_set_printerjob;
    int     HB_SET_GTMODE;
-   ULONG   HB_SET_BACKGROUNDTICK;
+   int     HB_SET_BACKGROUNDTICK;
    BOOL    HB_SET_HARDCOMMIT;
    BOOL    HB_SET_FORCEOPT;
    PHB_ITEM HB_SET_EOL;
